Added todo in App, extracted program Function in user to its own file
